Add files via upload
This code demonstrates several scenarios:

Global Context: When showThis is called, this refers to the global object (window in a browser). In strict mode, it would be undefined.

Object Method: Inside obj.method, this refers to the object obj because it's being called as a method of obj.

Constructor Call: When a function is used as a constructor (with the new keyword), this refers to the newly created object.

Arrow Functions: Arrow functions do not have their own this. Instead, they inherit this from the parent scope at the time they are defined. In the global scope, this will refer to the global object or undefined in strict mode, just like a regular function.

Explicit Binding: Using bind on showThis creates a new function where this is set to obj, no matter how the function is called.

Understanding these different contexts is essential for JavaScript developers, as it affects how functions interact with data and other parts of the program. ​